+++ desktop/src/main/kotlin/vn/loi/learning/desktop/ui/designsystem/LEIcons.kt
@@ -29,4 +29,6 @@ object LEIcons {
     val Keyboard: ImageVector = Icons.Default.Keyboard
     val Help: ImageVector = Icons.Default.HelpOutline
     val Settings: ImageVector = Icons.Default.Settings
+    val Undo: ImageVector = Icons.Default.Undo
+    val Pause: ImageVector = Icons.Default.Pause
 }